Fix fetching binary files
Treating responses in the `get` method as a String rather than a
Buffer can cause file contents to become corrupted. To fix this I have 
updated the `request` dependancy to 2.9.3 which includes handling of
responses as Buffer objects and modified the `get` method to explicitly 
use the buffer behaviour by setting encoding to null.